Change type in the internal structs and methods

+++ exonum/src/api/websocket.rs
@@ -114,7 +114,7 @@ impl CommittedTransactionSummary {
         let tx_result = schema.transaction_result(location)?;
         let location_proof = schema
             .block_transactions(location.block_height())
-            .get_proof(location.position_in_block());
+            .get_proof(location.position_in_block() as u64);
         let time = median_precommits_time(
             &schema
                 .block_and_precommits(location.block_height())
